refactor!: move qsystem functions to own module
BREAKING CHANGE: qsystem (formerly hseries) ops moved to dedicated module
