+++ tests/integration/test_tket.py
@@ -20,8 +20,9 @@ from guppylang.module import GuppyModule
 from guppylang.std.angles import pi
 from guppylang.std.builtins import owned, py
 from guppylang.std import quantum
+from guppylang.std.qsystem.functional import phased_x
 from guppylang.std.quantum import measure, qubit
-from guppylang.std.quantum_functional import phased_x, rz, zz_max
+from guppylang.std.quantum_functional import rz, zz_max
 from tests.util import guppy_to_circuit
 
 tket2_installed = find_spec("tket2") is not None
